feat: move more stable functions
